% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{check_xml_structure}
\alias{check_xml_structure}
\title{Check that two files have the same XML structure}
\usage{
check_xml_structure(path1, path2)
}
\arguments{
\item{path1}{First path}

\item{path2}{Second path}
}
\description{
Useful if \code{deepl_branch_update()} fails with an error about the XML structure.
Checkout the default branch, check the two files.
}
