% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languages.R
\name{deepl_languages}
\alias{deepl_languages}
\title{Languages supported by DeepL API}
\usage{
deepl_languages(type = c("target", "source"))
}
\arguments{
\item{type}{Either "target" or "source"}
}
\value{
A data.frame of languages (language code as "language",
name as "name", whether formality is supported as "supports_formality").
}
\description{
Languages supported by DeepL API
}
\examples{
\dontrun{
deepl_languages(type = "source")
deepl_languages(type = "target")
}

}
