% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{deepl_translate_clipboard}
\alias{deepl_translate_clipboard}
\title{Translate Markdown string from clipboard}
\usage{
deepl_translate_clipboard(
  target_lang,
  source_lang = NULL,
  glossary_name = NULL,
  formality = c("default", "more", "less", "prefer_more", "prefer_less")
)
}
\arguments{
\item{target_lang}{Name or code of source language. See \href{https://www.deepl.com/docs-api/general/get-languages/}{DeepL docs}.}

\item{source_lang}{Name or code of source language. Guessed using cld3 if not provided.
See \href{https://www.deepl.com/docs-api/general/get-languages/}{DeepL docs}.}

\item{glossary_name}{Name of the glossary to be used, if any (character).}

\item{formality}{Formality level to use (character), one of
\itemize{
\item "default" (default)
\item "less" -- for a more informal language
\item "prefer_more" -- for a more formal language if available, otherwise fallback to default formality
\item "prefer_less" -- for a more informal language if available, otherwise fallback to default formality
}}
}
\value{
Translated Markdown string, also put on the clipboard.
}
\description{
Read the string from the clipboard using the clipr package.
Guesses the source language using the cld3 package.
}
\details{
To make this function less chatty, set the \code{babeldown.quiet}
option to \code{TRUE}.
}
\examples{
\dontrun{
clipr::write_clip("Je suis en vacances, en fait.")
deepl_translate_clipboard(target_lang = "EN-US")

clipr::write_clip("Je suis en vacances.")
deepl_translate_clipboard(target_lang = "EN-US")
}
}
