% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate-hugo.R
\name{deepl_translate_hugo}
\alias{deepl_translate_hugo}
\title{Translate a Hugo post}
\usage{
deepl_translate_hugo(
  post_path = NULL,
  force = FALSE,
  yaml_fields = c("title", "description"),
  glossary_name = NULL,
  source_lang = NULL,
  target_lang = NULL,
  formality = c("default", "more", "less", "prefer_more", "prefer_less")
)
}
\arguments{
\item{post_path}{Path to post. If RStudio IDE is installed, it will default to the
currently open document.
If you use Quarto or R Markdown for Hugo, translate the source file (qmd or Rmd)
and then render it to Markdown.}

\item{force}{Whether to overwrite the post in the target language.}

\item{yaml_fields}{Vector of character names of YAML fields to translate.}

\item{glossary_name}{Name of the glossary to be used, if any (character).}

\item{source_lang}{Name or code of source language. See \href{https://www.deepl.com/docs-api/general/get-languages/}{DeepL docs}.}

\item{target_lang}{Name or code of source language. See \href{https://www.deepl.com/docs-api/general/get-languages/}{DeepL docs}.}

\item{formality}{Formality level to use (character), one of
\itemize{
\item "default" (default)
\item "less" -- for a more informal language
\item "prefer_more" -- for a more formal language if available, otherwise fallback to default formality
\item "prefer_less" -- for a more informal language if available, otherwise fallback to default formality
}}
}
\value{
None
}
\description{
This assumes the Hugo website uses
\itemize{
\item leaf bundles (each post in a folder, \code{leaf-bundle/index.md});
\item multilingualism so that a post in say Spanish lives in \code{leaf-bundle/index.es.md}.
}
}
\examples{
\dontrun{
dir <- withr::local_tempdir()
blogdown::new_site(dir = dir)
deepl_translate_hugo(
  file.path(dir, "content", "post", "2016-12-30-hello-markdown", "index.md"),
  source_lang = "en",
  target_lang = "fr",
  formality = "less"
)
readLines(file.path(dir, "content", "post", "2016-12-30-hello-markdown", "index.fr.md"))
}
}
