% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{deepl_translate_markdown_string}
\alias{deepl_translate_markdown_string}
\title{Translate Markdown string}
\usage{
deepl_translate_markdown_string(
  markdown_string,
  glossary_name = NULL,
  source_lang,
  target_lang,
  formality = c("default", "more", "less", "prefer_more", "prefer_less")
)
}
\arguments{
\item{markdown_string}{Markdown string to translate}

\item{glossary_name}{Name of the glossary to be used, if any (character).}

\item{source_lang}{Name or code of source language. See \href{https://www.deepl.com/docs-api/general/get-languages/}{DeepL docs}.}

\item{target_lang}{Name or code of source language. See \href{https://www.deepl.com/docs-api/general/get-languages/}{DeepL docs}.}

\item{formality}{Formality level to use (character), one of
\itemize{
\item "default" (default)
\item "less" -- for a more informal language
\item "prefer_more" -- for a more formal language if available, otherwise fallback to default formality
\item "prefer_less" -- for a more informal language if available, otherwise fallback to default formality
}}
}
\value{
Translated Markdown string
}
\description{
Translate Markdown string
}
\examples{
\dontrun{
deepl_translate_markdown_string(
  "[So _incredibly_ **wonderful**](https://ropensci.org)!",
  source_lang = "EN",
  target_lang = "FR",
  formality = "less"
)
}
}
