% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate-quarto.R
\name{deepl_translate_quarto}
\alias{deepl_translate_quarto}
\title{Translate a Quarto book chapter}
\usage{
deepl_translate_quarto(
  book_path,
  chapter,
  force = FALSE,
  render = TRUE,
  yaml_fields = c("title", "description"),
  glossary_name = NULL,
  source_lang = NULL,
  target_lang = NULL,
  formality = c("default", "more", "less", "prefer_more", "prefer_less")
)
}
\arguments{
\item{book_path}{Path to book source}

\item{chapter}{Chapter name}

\item{force}{Whether to overwrite the chapter in the target language.}

\item{render}{Whether to run \code{babelquarto::render_bool()} after translation.}

\item{yaml_fields}{Vector of character names of YAML fields to translate.}

\item{glossary_name}{Name of the glossary to be used, if any (character).}

\item{source_lang}{Name or code of source language. See \href{https://www.deepl.com/docs-api/general/get-languages/}{DeepL docs}.}

\item{target_lang}{Name or code of source language. See \href{https://www.deepl.com/docs-api/general/get-languages/}{DeepL docs}.}

\item{formality}{Formality level to use (character), one of
\itemize{
\item "default" (default)
\item "less" -- for a more informal language
\item "prefer_more" -- for a more formal language if available, otherwise fallback to default formality
\item "prefer_less" -- for a more informal language if available, otherwise fallback to default formality
}}
}
\value{
None
}
\description{
This assumes the book is set up à la \href{https://github.com/ropensci-review-tools/babelquarto}{babelquarto}.
}
\examples{
\dontrun{
temp_dir <- withr::local_tempdir()
babelquarto::quarto_multilingual_book(
  parent_dir = temp_dir,
  book_dir = "blop",
  main_language = "en",
  further_languages = "es"
)
book_path <- file.path(temp_dir, "blop")
deepl_translate_quarto(
  book_path = book_path,
  chapter = "intro.qmd",
  force = TRUE, # the existing chapter is a placeholder
  render = TRUE,
  source_lang = "EN",
  target_lang = "ES",
  formality = "less"
)
# have a look at the translation
readLines(file.path(book_path, "intro.es.qmd"))
# servr::httw(file.path(book_path, "_book"))
}
}
