% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto-multilingual-book-render.R
\name{render_quarto_multilingual_book}
\alias{render_quarto_multilingual_book}
\title{Render a Quarto multilingual book}
\usage{
render_quarto_multilingual_book(book_folder)
}
\arguments{
\item{book_folder}{Path where the book source is located}

\item{output_folder}{Path of the folder where the book will be rendered}
}
\description{
Render a Quarto multilingual book
}
\details{
babeldown expects a book folder with
each qmd/Rmd present in as many languages as needed,
with the same basename but,
\itemize{
\item once with only \code{.qmd} as extension for the main language,
\item once with \code{.es.qmd} (using the language code) for each other language.
}

You also need to register the language in the configuration file:

\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{babeldown:
  mainlanguage: 'en'
  languages: ['es', 'fr']
}\if{html}{\out{</div>}}
}
\examples{
directory <- withr::local_tempdir()
babeldown:::quarto_multilingual_book(directory, subdir = "blop")
render_quarto_multilingual_book(file.path(directory, "blop"))
if (require("servr") && rlang::is_interactive()) {
  servr::httw(file.path(directory, "blop", "_book"))
}

}
