% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary.R
\name{upsert_glossary}
\alias{upsert_glossary}
\title{Create or update glossary}
\usage{
upsert_glossary(filename, glossary_name = NULL, source_lang, target_lang)
}
\arguments{
\item{filename}{Name of the glossary file.
See \href{https://www.deepl.com/en/docs-api/glossaries/formats/}{DeepL docs on supported formats}.
babeldown is stricter: the columns need to be named like \code{source_lang} and \code{target_lang}.}

\item{glossary_name}{Name for the glossary. Defaults to the filename without extension.}

\item{source_lang}{Name or code of source language. See \href{https://www.deepl.com/docs-api/general/get-languages/}{DeepL docs}.}

\item{target_lang}{Name or code of source language. See \href{https://www.deepl.com/docs-api/general/get-languages/}{DeepL docs}.}
}
\value{
glossary ID
}
\description{
Create or update glossary
}
\examples{
\dontrun{
upsert_glossary(
  system.file("example-es-en.csv", package = "babeldown"),
  glossary_name = "rstats-glosario",
  target_lang = "Spanish",
  source_lang = "English"
)
}
}
