% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{deepl_translate}
\alias{deepl_translate}
\title{Translate a Markdown file}
\usage{
deepl_translate(
  path,
  out_path,
  glossary_name = NULL,
  source_lang = NULL,
  target_lang = NULL,
  formality = c("default", "more", "less", "prefer_more", "prefer_less")
)
}
\arguments{
\item{path}{Path to the Markdown file to be translated (character).}

\item{out_path}{Path where the new translated file should be saved (character).}

\item{glossary_name}{Name of the glossary to be used, if any (character).}

\item{source_lang}{Name or code of source language. See \href{https://www.deepl.com/docs-api/general/get-languages/}{DeepL docs}.}

\item{target_lang}{Name or code of source language. See \href{https://www.deepl.com/docs-api/general/get-languages/}{DeepL docs}.}

\item{formality}{Formality level to use (character), one of
\itemize{
\item "default" (default)
\item "more" -- for a more formal language
\item "less" -- for a more informal language
\item "prefer_more" -- for a more formal language if available, otherwise fallback to default formality
\item "prefer_less" -- for a more informal language if available, otherwise fallback to default formality
}}
}
\value{
None
}
\description{
Translate a Markdown file
}
