% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{deepl_update}
\alias{deepl_update}
\alias{deepl_branch_update}
\title{Update a translation of a file in a Git repo}
\usage{
deepl_update(
  path,
  out_path,
  yaml_fields = c("title", "description"),
  glossary_name = NULL,
  source_lang = NULL,
  target_lang = NULL,
  formality = c("default", "more", "less", "prefer_more", "prefer_less"),
  max_commits = 100L
)

deepl_branch_update(repo = ".", max = 100)
}
\arguments{
\item{path}{Path to the Markdown file to be translated (character).}

\item{out_path}{Path where the new translated file should be saved (character).}

\item{yaml_fields}{Vector of character names of YAML fields to translate.}

\item{glossary_name}{Name of the glossary to be used, if any (character).}

\item{source_lang}{Name or code of source language. See \href{https://www.deepl.com/docs-api/general/get-languages/}{DeepL docs}.}

\item{target_lang}{Name or code of source language. See \href{https://www.deepl.com/docs-api/general/get-languages/}{DeepL docs}.}

\item{formality}{Formality level to use (character), one of
\itemize{
\item "default" (default)
\item "less" -- for a more informal language
\item "prefer_more" -- for a more formal language if available, otherwise fallback to default formality
\item "prefer_less" -- for a more informal language if available, otherwise fallback to default formality
}}

\item{max_commits}{Maximal number of commits to go back to to find
when the target and source files were updated.
You might need to increase it if your project is very active.}

\item{repo}{path to the Git repository.}
}
\value{
None
}
\description{
Re-use existing translation where possible
(at the node level: paragraph, heading, etc.)
\code{deepl_update()} is based on the commit history in a single branch,
\code{deepl_branch_update()} is based on the commit history in a branch
against another (in a GitHub PR for instance), and requires a
configuration file (see "Configuration" section).
}
\section{Markdown formatting, both \code{deepl_update()} and \code{deepl_branch_update()}}{

The update functions will re-use translations as much as possible, at the node level.
If the files diverged a bit from the tinkr Markdown formatting, you might notice other
edits that change the Markdown formatting.
You could choose not to stage them, or you could do a roundtrip with tinkrin all files
(all language versions) prior to
any other edits. \verb{tinkr::yarn$new(<filename>)$write(<filename>)}.
}

\section{Limitations of \code{deepl_update()}}{

\code{deepl_update()} looks for the latest commit that updated the source file,
and for the latest commit that updated the target file.
If the target file was updated later than the source file,
or at the same time,
nothing happens: you might need to
fix up the Git history with rebase for instance.
Also note that if there were some commits irrelevant to the translation
but containing edits to the source file, like typo fixes,
too much of the target file will be translated automatically again.
}

\section{Configuration for \code{deepl_branch_update()}}{

\code{deepl_branch_update()} is more automatic but needs configuration.
In a branch, for any edits made to any Rmd/qmd/md file, it will update
the translations of the Rmd/qmd/md file with the same name and a language code.
For instance if a branch had commits updating \code{pkg_building.Rmd},
\code{deepl_branch_update()} will update \code{pkg_building.es.Rmd}.
Do not run it if your PR updated both \code{pkg_building.Rmd} and
\code{pkg_building.es.Rmd}.
Make sure to read the section about Markdown formatting.

The configuration needed must live under \verb{_deepl.yml} at the root of the Git repo.
It must contains those fields:
\itemize{
\item "preferences" that for any source-target pair, indicates the preference for the
\code{formality}, \code{glossary} and \code{yaml_fields} arguments of \code{deepl_translate()}.
\item "excludes" that indicates which files to exclude from the automatic detection
(all Rmd/qmd/md files are otherwise considered). It is used as \code{glob} arguments in
calls to \code{fs::dir_ls()}.
\item "languages" that for any language code used in file names, indicates what
DeepL API code to use when it is a source or target language.
}

Example (\verb{_deepl.yml}):

\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{preferences:
  - source: en
    target: es-419
    formality: less
    glossary: "glosario"
    yaml_fields: ["title"]
  - source: en
    target: pt-br
    formality: less
    yaml_fields: ["title", "description"]
  - source: pt
    target: en-us
    formality: prefer_less
    yaml_fields: ["title", "description"]
  - source: es
    target: en-us
    formality: prefer_less
    yaml_fields: ["title", "description"]

excludes: ["*.md"]

languages:
  - extension: es
    target: es-419
    source: es
  - extension: ''
    target: en-us
    source: en
  - extension: pt
    target: pt-br
    source: pt
    

}\if{html}{\out{</div>}}
}

